+++ test/special_teams_test.js
@@ -50,6 +50,7 @@ function driveToSpecial(c, guardMax) {
         else if (p.kind === 'kickoff') { C.callKickoff(c, C.kickoffChoices(c).recommendation); }
         else if (p.kind === 'pat') { C.callPat(c, C.patChoices(c).recommendation); }
         else if (p.kind === 'defspecial') { C.callDefSpecial(c, C.defSpecialChoices(c).recommendation); }
+        else if (p.kind === 'penalty') { C.callPenalty(c, C.penaltyChoices(c).recommendation); }
         else if (p.kind === 'auto') { C.advance(c); }
         C.reports(c);
     }
@@ -123,6 +124,7 @@ module.exports = function (t) {
         else if (pp.kind === 'kickoff') { C.callKickoff(opp, C.kickoffChoices(opp).recommendation); }
         else if (pp.kind === 'pat') { C.callPat(opp, C.patChoices(opp).recommendation); }
         else if (pp.kind === 'defspecial') { C.callDefSpecial(opp, C.defSpecialChoices(opp).recommendation); }
+        else if (pp.kind === 'penalty') { C.callPenalty(opp, C.penaltyChoices(opp).recommendation); }
         else if (pp.kind === 'auto') { C.advance(opp); }
         C.reports(opp);
     }